Just base stats instead of MobStats component

+++ src/gamestate.rs
@@ -49,7 +49,7 @@ impl GameState {
 
         // Draw heart containers.
         for i in range(0, (max_hp + 1) / 2) {
-            let pos = V2(i * 8, 8);
+            let pos = V2(i as int * 8, 8);
             let idx = if hp >= (i + 1) * 2 { icon::HEART }
                 else if hp == i * 2 + 1 { icon::HALF_HEART }
                 else { icon::NO_HEART };
